update controllers v2

+++ ReadingComicWebSite/src/main/java/com/example/readingcomicwebsite/service/ICommentService.java
@@ -6,7 +6,8 @@ import java.util.List;
 
 public interface ICommentService {
     List<Comment> findAll();
-    List<Comment> findAllByMangaId(Long mangaId);
+
+    List<Comment> findAllByMangaId(Integer mangaId);
 
     Comment save(Comment comment);
